+++ frontend/src/pages/RollingpaperPage/hooks/useMessageForm.tsx
@@ -5,7 +5,7 @@ import useCheckBox from "@/hooks/useCheckBox";
 
 const INIT_COLOR = COLORS.YELLOW;
 
-interface UseUpdateMessageProp {
+interface UseMessageFormArgs {
   initContent?: string;
   initColor?: string;
   initAnonymous?: boolean;
@@ -17,7 +17,7 @@ const useMessageForm = ({
   initColor = INIT_COLOR,
   initAnonymous = false,
   initSecret = false,
-}: UseUpdateMessageProp) => {
+}: UseMessageFormArgs) => {
   const [content, setContent] = useState(initContent);
   const [color, setColor] = useState(initColor);
